Move lib files to respective agent, controller and common directories
