+++ agent/main.py
@@ -7,11 +7,11 @@ from opentelemetry import trace
 
 from agent import config, task_api, tracing
 from agent.executors import get_executor_api
+from agent.lib.docker import docker, get_network_config_args
 from common import config as common_config
 from common.job_executor import ExecutorAPI, ExecutorState, JobDefinition, JobStatus
+from common.lib.log_utils import configure_logging, set_log_context
 from common.schema import TaskType
-from jobrunner.lib.docker import docker, get_network_config_args
-from jobrunner.lib.log_utils import configure_logging, set_log_context
 
 
 log = logging.getLogger(__name__)
