+++ common/lib/string_utils.py
