+++ controller/cli/add_job.py
@@ -11,14 +11,14 @@ import textwrap
 from pathlib import Path
 from urllib.parse import urlparse
 
+from common.lib.git import get_sha_from_remote_ref
+from common.lib.log_utils import configure_logging
 from controller.cli.utils import add_backend_argument
 from controller.create_or_update_jobs import create_or_update_jobs
+from controller.lib.database import find_where
 from controller.models import Job, random_id
 from controller.sync import job_request_from_remote_format
 from jobrunner import tracing
-from jobrunner.lib.database import find_where
-from jobrunner.lib.git import get_sha_from_remote_ref
-from jobrunner.lib.log_utils import configure_logging
 
 
 def main(
