+++ controller/cli/prepare_for_reboot.py
@@ -6,10 +6,10 @@ automatically re-run after a reboot.
 import argparse
 
 from controller.cli.utils import add_backend_argument
+from controller.lib.database import find_where, transaction
 from controller.main import cancel_job, set_code
 from controller.models import Job, State, StatusCode, Task, TaskType
 from controller.queries import get_flag_value
-from jobrunner.lib.database import find_where, transaction
 
 
 def main(backend, status=False, require_confirmation=True):
