+++ controller/create_or_update_jobs.py
@@ -14,8 +14,15 @@ import time
 from pipeline import RUN_ALL_COMMAND, ProjectValidationError, load_pipeline
 
 from common import config as common_config
+from common.lib.git import GitError, GitFileNotFoundError, read_file_from_repo
+from common.lib.github_validators import (
+    GithubValidationError,
+    validate_branch_and_commit,
+    validate_repo_url,
+)
 from controller import config
 from controller.actions import get_action_specification
+from controller.lib.database import exists_where, insert, transaction, update_where
 from controller.models import Job, SavedJobRequest, State, StatusCode
 from controller.queries import calculate_workspace_state
 from controller.reusable_actions import (
@@ -23,13 +30,6 @@ from controller.reusable_actions import (
     resolve_reusable_action_references,
 )
 from jobrunner import tracing
-from jobrunner.lib.database import exists_where, insert, transaction, update_where
-from jobrunner.lib.git import GitError, GitFileNotFoundError, read_file_from_repo
-from jobrunner.lib.github_validators import (
-    GithubValidationError,
-    validate_branch_and_commit,
-    validate_repo_url,
-)
 
 
 log = logging.getLogger(__name__)
