+++ jobrunner/tracing.py
@@ -12,9 +12,9 @@ from opentelemetry.trace import propagation
 from opentelemetry.trace.propagation.tracecontext import TraceContextTextMapPropagator
 
 from common import config as common_config
+from common.lib import warn_assertions
 from controller.models import Job, State, StatusCode
 from controller.queries import get_saved_job_request
-from jobrunner.lib import warn_assertions
 
 
 logger = logging.getLogger(__name__)
