+++ tests/agent/lib/test_docker_stats.py
