+++ tests/agent/test_agent.py
@@ -8,8 +8,8 @@ import pytest
 from agent import config, main, task_api
 from common.job_executor import ExecutorState, JobDefinition
 from controller import task_api as controller_task_api
+from controller.lib.database import update_where
 from controller.models import Task, TaskType
-from jobrunner.lib.database import update_where
 from tests.agent.stubs import StubExecutorAPI
 from tests.conftest import get_trace
