+++ tests/common/lib/test_string_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from jobrunner.lib.string_utils import project_name_from_url, slugify, tabulate
+from common.lib.string_utils import project_name_from_url, slugify, tabulate
 
 
 @pytest.mark.parametrize(
