+++ tests/controller/cli/test_prepare_for_reboot.py
@@ -1,9 +1,9 @@
 import pytest
 
 from controller.cli import prepare_for_reboot
+from controller.lib import database
 from controller.models import Job, State, StatusCode, Task, TaskType
 from controller.queries import set_flag
-from jobrunner.lib import database
 from tests.conftest import get_trace
 from tests.factories import (
     canceljob_db_task_factory,
