+++ tests/controller/lib/test_database.py
@@ -167,7 +167,7 @@ def test_exists_where(tmp_work_dir):
     assert job_id_exists is True
 
 
-def exists_where_null(tmp_work_dir):
+def test_exists_where_null(tmp_work_dir):
     insert(Job(id="foo123", backend=None))
     insert(Job(id="foo124", backend="test"))
     insert(Job(id="foo125", backend="foo"))
