+++ tests/controller/webapp/test_management.py
@@ -7,8 +7,8 @@ that they run correctly via management command.
 from django.core.management import call_command
 
 from controller import queries
+from controller.lib import database
 from controller.models import Job, State, StatusCode, Task, TaskType
-from jobrunner.lib import database
 from tests.controller.cli.test_flags import TEST_DATESTR, TEST_TIME
 from tests.controller.cli.test_prepare_for_reboot import pause_backend
 from tests.factories import job_factory, runjob_db_task_factory
